Cleanup code style

+++ src/Utils/SysCalls.cs
@@ -18,24 +18,20 @@ namespace trakr_sharp.Utils {
         private static readonly string _screenshotsPath = "user_data/screenshots";
         private static readonly string _readmePath = "README.md";
 
-        private static readonly string[] _ignoredProcs = new string[] {"System Idle Process", "explorer.exe", "smss.exe",
+        private static readonly string[] _ignoredProcs = new string[] { "System Idle Process", "explorer.exe", "smss.exe",
             "taskmgr.exe", "spoolsv.exe", "lsass.exe", "csrss.exe", "winlogon.exe", "svchost.exe", "System",
             "Registry", "WindowsInternal.ComposableShell.Experiences.TextInput.InputApp.exe", "conhost.exe" };
 
-        // Creates a json to store user settings if it doens't exist
+        /// <summary>
+        /// Creates a JSON to store user settings if it doens't exist
+        /// </summary>
         public static void InitUserSettings() {
             // Create user_data folder if not exists
             Directory.CreateDirectory("user_data");
 
             // Create settings.json if not exists
             if (!File.Exists(_settingsJsonPath)) {
-                UserSettings userSettings = new UserSettings {
-                    OnClose = UserSettings.CloseBehaviour.Close,
-                    RunOnStartup = false,
-                    EnableScreenshots = false,
-                    ShowUtilCols = false
-                };
-
+                UserSettings userSettings = UserSettings.Default;
                 // Convert userSettings to json string and write to "user_data/settings.json"
                 using (StreamWriter file = File.CreateText(_settingsJsonPath)) {
                     JsonSerializer serializer = new JsonSerializer();
@@ -44,7 +40,9 @@ namespace trakr_sharp.Utils {
             }
         }
 
-        // Reads "user_data/settings.json" and returns a UserSettings object
+        /// <summary>
+        /// Reads "user_data/settings.json" and returns a UserSettings object
+        /// </summary>
         public static UserSettings ReadUserSettings() {
             UserSettings userSettings;
 
@@ -57,7 +55,9 @@ namespace trakr_sharp.Utils {
             return userSettings;
         }
 
-        // Writes changes to "user_data/settings.json"
+        /// <summary>
+        /// Writes the provided UserSettings obj to "user_data/settings.json"
+        /// </summary>
         public static void UpdateUserSettings(UserSettings newUserSettings) {
             // Read "user_data/settings/json" and save deserialzied obj in jsonObj
             string diskString = File.ReadAllText(_settingsJsonPath);
@@ -74,11 +74,17 @@ namespace trakr_sharp.Utils {
             File.WriteAllText(_settingsJsonPath, outputJson);
         }
 
-        public static void InitScreenshots() {
-            // Create user_data folder if not exists
+        /// <summary>
+        /// Creates the user_data folder if it doesn't exist
+        /// </summary>
+        public static void InitScreenshotsDir() {
             Directory.CreateDirectory(_screenshotsPath);
         }
 
+        /// <summary>
+        /// Attempts to take and saves screenshot of the current display
+        /// </summary>
+        /// <returns>[String] The path the image was saved to, or a failure message</returns>
         public static string TakeScreenshot() {
             string fileName = Guid.NewGuid().ToString() + ".png";
             string filePath = string.Format("{0}/{1}", _screenshotsPath, fileName);
@@ -106,7 +112,9 @@ namespace trakr_sharp.Utils {
             }
         }
 
-        // Opens the "README.md" file with notepad
+        /// <summary>
+        /// Opens the "README.md" file with notepad
+        /// </summary>
         public static bool OpenReadme() {
             try {
                 Process.Start("notepad.exe", _readmePath);
@@ -117,7 +125,9 @@ namespace trakr_sharp.Utils {
             }
         }
 
-        // Opens the trakr-sharp repo link in a browser window
+        /// <summary>
+        /// Opens the trakr-sharp repo link in a browser window
+        /// </summary>
         public static bool OpenRepoLink() {
             try {
                 Process.Start("https://github.com/bluetayden/trakr-sharp");
@@ -128,43 +138,42 @@ namespace trakr_sharp.Utils {
             }
         }
 
-        // List of currently running processes on the device
-        public static List<string> GetRunningProcNameList(bool incDuplicates = false) {
+        /// <summary>
+        /// Returns a list of processes currently running on the device
+        /// </summary>
+        public static List<string> GetRunningProcNameList(bool distinct = true) {
             // Get list of all running process names
             IEnumerable<string> runningProcs = Process.GetProcesses().Select(proc => proc.ProcessName + ".exe");
-
-            // Filter unwanted processes from list
+            // Filter unwanted processes (and duplicates if requested)
             IEnumerable<string> filteredProcs = runningProcs.Where(proc => !_ignoredProcs.Contains(proc));
-            if (!incDuplicates) {
-                // Remove duplicates from list
-                filteredProcs = filteredProcs.Distinct();
-            }
+            if (distinct) { filteredProcs = filteredProcs.Distinct(); }
             // Order list alphabetically
             filteredProcs = filteredProcs.OrderBy(proc => proc[0]);
 
             return filteredProcs.ToList();
         }
 
-        // List of currently running processes on the device, without procs that are already stored in the db
+        /// <summary>
+        /// Returns a list of running processes excluding ones that are already stored in the db
+        /// </summary>
         public static List<string> GetRunningUntrackedProcNameList() {
-            // Get list of all running process names
+            // Get list of all running process names (append .exe to each name)
             IEnumerable<string> runningProcs = Process.GetProcesses().Select(proc => proc.ProcessName + ".exe");
-
             // Filter unwanted processes from list
             IEnumerable<string> filteredProcs = runningProcs.Where(proc => !_ignoredProcs.Contains(proc));
-            // Get list of processes stored in database
+            // Get list of processes stored in db and filter stored procs out
             List<string> storedProcs = Database.GetProcessNameList();
-            // Filter stored procs from list
             filteredProcs = filteredProcs.Where(proc => !storedProcs.Contains(proc));
-            // Remove duplicates from list
+            // Remove duplicates and order alphabetically
             filteredProcs = filteredProcs.Distinct();
-            // Order list alphabetically
             filteredProcs = filteredProcs.OrderBy(proc => proc[0]);
 
             return filteredProcs.ToList();
         }
 
-        // Return a dict that contains paths to exe files with proc_names as keys. (Uses selectedProcs as a predicate).
+        /// <summary>
+        /// Returns a dict that contains paths to exe files with proc_names as keys. (Uses selectedProcs as a predicate).
+        /// </summary>
         public static Dictionary<string, string> GetProcPathPairs(List<string> selectedProcs) {
             Dictionary<string, string> procPathPairs = new Dictionary<string, string>();
 
@@ -186,30 +195,28 @@ namespace trakr_sharp.Utils {
                     procPathPairs.Add(item.Name, item.Path);
                 }
             }
-
             // Add items that were not found by WMI (process likely closed while AddRecordsForm was open)
             foreach (string proc_name in selectedProcs) {
                 if (!procPathPairs.ContainsKey(proc_name)) {
                     procPathPairs.Add(proc_name, "");
                 }
             }
-
             mo_searcher.Dispose(); // Manual disposal required to stop WMI mem leaks
             mo_results.Dispose();
 
             return procPathPairs;
         }
 
-        // List of <string name, int count> of currently running process on the device that are being tracked in the db
+        /// <summary>
+        /// List of [string name, int count] of currently running process on the device, that are also being tracked in the db
+        /// </summary>
         public static Dictionary<string, int> GetRunningTrackedPairs() {
             List<string> trackedProcs = Utils.Database.GetProcessNameList();
-
             // Filter currently running process to only contain those being tracked (include duplicates)
             trackedProcs = GetRunningProcNameList(true).Where(proc => trackedProcs.Contains(proc)).ToList();
 
             // Store name keys in trackedPairs, and inc count for every ocurrence of a particular proc
             Dictionary<string, int> trackedPairs = new Dictionary<string, int>();
-
             foreach (string p in trackedProcs) {
                 // If a key == p in trackedPairs, inc count
                 if (trackedPairs.ContainsKey(p)) {
@@ -220,7 +227,6 @@ namespace trakr_sharp.Utils {
                     trackedPairs.Add(p, 1);
                 }
             }
-
             return trackedPairs;
         }
 
@@ -244,12 +250,12 @@ namespace trakr_sharp.Utils {
             foreach (string name in procPathPairs.Keys) {
                 string imgPath = String.Format("cache/{0}.png", name.Substring(0, name.Length - 4));
 
-                // If icon not already cached, or overwrite is true
+                // If icon not already cached, or overwriting is enabled
                 if (!System.IO.File.Exists(imgPath) || overwrite) {
                     // Get icon from path
                     using (Icon ico = GetIconFromPath(procPathPairs[name])) {
                         // If the icon does already exist in the cache, delete it to prevent GDI error
-                        if (System.IO.File.Exists(imgPath)) {
+                        if (overwrite && System.IO.File.Exists(imgPath)) {
                             System.IO.File.Delete(imgPath);
                         }
 
@@ -262,7 +268,9 @@ namespace trakr_sharp.Utils {
             }
         }
 
-        // Retrieves an image from the cache folder using a proc_name
+        /// <summary>
+        /// Retrieves an image from the cache folder using a proc_name
+        /// </summary>
         public static Bitmap GetIconFromCache(string name) {
             Bitmap retImg;
